Remove static IP and use host network.
This is a big change and will result in a minor version bump. Removing the
assumption on IP addresses will make most of the setup quite easy and remove
all the hardcoded IPs.

The places which necessarily do require one can be based of 127.0.0.1, like
MAILMAN_HOST in Hyperkitty, which verifies the sender based on the incoming IP
address.