Add lynx to the mailman core Docker image
Mailman's convert_html_to_plaintext setting relies on the value of
html_to_plain_text_command which is by default set to:

  /usr/bin/lynx -dump $filename

This will fail on the current image because lynx is not installed.
Simply add lynx to the list of packages to install to correct this.
The file path for the lynx command installed via the Alpine package does
seem to be /usr/bin/lynx which means we don't need to update any
configuration to use this command. The defaults are sufficient.