Initial commit of pod creation
* Added script `./generate_pod.sh`, run it to generate a pod.

This script should create a functioning mailman3 instance which you can,

* Access at http://localhost:8000/
* Generate a kubernetes yaml with `podman generate kube mailman`

This works functionally similar to docker-compose, except everything
will run rootless, and you don't need the docker daemon. It also helps
unify the deployment as a single pod on a single machine, with potential
deployment of a pod on kubernetes.