[web, postorius] Fix libldap installation package group
The libldap package is not installed in the mailman-web:0.5 and
postorius images despite being listed as a dependency in the Dockerfile.
This issue arises because libldap is included in the .build-deps virtual
package group, which is removed at the end of the build process, causing
the package to be uninstalled.

This commit addresses the issue by moving libldap from the .build-deps
virtual package group to the .mailman-rundeps virtual package group in
both the web and postorius Dockerfiles. This ensures that libldap
remains installed in the final image, as it is now part of the runtime
dependencies.

Changes:
- Move libldap installation from .build-deps to .mailman-rundeps in
  postorius/Dockerfile and postorius/Dockerfile.env;
- Move libldap installation from .build-deps to .mailman-rundeps in
  web/Dockerfile and web/Dockerfile.env.

These changes are necessary to ensure that the libldap package is
available in the running containers, preventing runtime errors related
to missing LDAP dependencies.

Signed-off-by: Antonio Rocco <8lue@8lue.xyz>