Create the var directory before trying to chown.
During cold start, var directory may not exist and we try to chown
the var directory which fails if it doesn't exist. We want to make
and chown the var directory in that case otherwise, mailman will not
be able to write to path `/opt/mailman` to create it's own var
directory.